+++ apps/domain/test/entity/Post.test.ts
@@ -1,4 +1,4 @@
-import { DIFFICULTY_KEY, USER_STATUS_KEY } from '@recipeaceful/library/dist/const'
+import { DIFFICULTY_KEY } from '@recipeaceful/library/dist/const'
 import { Post } from '../../src/entity/Post'
 import { ulid } from 'ulid'
 import { Calories } from '../../src/valueObject/Calories'
@@ -6,9 +6,8 @@ import { Difficulty } from '../../src/valueObject/Difficulty'
 import { Material } from '../../src/valueObject/Material'
 import { PostDetail } from '../../src/valueObject/PostDetail'
 import { PostTitle } from '../../src/valueObject/PostTitle'
-import { ProcessDetail } from '../../src/valueObject/ProcessDetail'
 import { PostId, UserId } from '../../src/valueObject/Ulid'
-import { UserStatus } from '../../src/valueObject/UserStatus'
+import { PostImage } from 'valueObject/PostImage'
 
 describe('Post', () => {
   const postId = new PostId(ulid())
@@ -19,6 +18,7 @@ describe('Post', () => {
         postId: postId,
         title: new PostTitle('卵焼き'),
         detail: new PostDetail('ちょー簡単な料理です。'),
+        images: null,
         calories: new Calories(120),
         difficulty: new Difficulty(DIFFICULTY_KEY.EASY),
         createdAt: new Date(),
@@ -32,18 +32,7 @@ describe('Post', () => {
             count: 1
           }
         ],
-        processes: [
-          {
-            detail: new ProcessDetail('フライパンを熱して卵を入れます。'),
-            image: 'material.png'
-          },
-          {
-            detail: new ProcessDetail('ちょっとしたらフライパンからお皿にのっけて終わり！'),
-            image: null
-          }
-        ],
         userId: userId,
-        userStatus: new UserStatus(USER_STATUS_KEY.ACTIVE),
         likes: [new UserId(ulid()), new UserId(ulid()), new UserId(ulid()), new UserId(ulid())]
       })
 
@@ -52,16 +41,11 @@ describe('Post', () => {
       expect(entity.detail.get()).toBe('ちょー簡単な料理です。')
       expect(entity.calories.get()).toBe(120)
       expect(entity.difficulty.get()).toBe(DIFFICULTY_KEY.EASY)
-      entity.materials.forEach((material, i) => {
-        expect(material.count).toBe(entity.materials[i]?.count)
-        expect(material.name.get()).toBe(entity.materials[i]?.name.get())
-      })
-      entity.processes.forEach((process, i) => {
-        expect(process.detail.get()).toBe(entity.processes[i]?.detail.get())
-        expect(process.image).toBe(entity.processes[i]?.image)
+      entity.materials?.forEach((material, i) => {
+        expect(material.count).toBe(entity.materials![i]?.count)
+        expect(material.name.get()).toBe(entity.materials![i]?.name.get())
       })
       expect(entity.userId.get()).toBe(userId.get())
-      expect(entity.userStatus.get()).toBe(USER_STATUS_KEY.ACTIVE)
       entity.likes?.forEach((like, i) => {
         expect(like.get()).toBe(entity.likes![i]?.get())
       })
@@ -73,6 +57,7 @@ describe('Post', () => {
         title: new PostTitle('卵焼き'),
         detail: new PostDetail('ちょー簡単な料理です。'),
         calories: new Calories(120),
+        images: null,
         difficulty: new Difficulty(DIFFICULTY_KEY.EASY),
         createdAt: new Date(),
         materials: [
@@ -81,14 +66,7 @@ describe('Post', () => {
             count: 1
           }
         ],
-        processes: [
-          {
-            detail: new ProcessDetail('フライパンを熱して卵を入れます。'),
-            image: 'material.png'
-          }
-        ],
         userId: userId,
-        userStatus: new UserStatus(USER_STATUS_KEY.ACTIVE),
         likes: null
       })
 
@@ -97,46 +75,25 @@ describe('Post', () => {
     })
   })
   describe('error cases', () => {
-    it('If the processes does not exist', () => {
+    it('画像が６枚以上', () => {
       expect(() =>
         Post.create({
           postId: postId,
           title: new PostTitle('卵焼き'),
           detail: new PostDetail('ちょー簡単な料理です。'),
-          calories: new Calories(120),
-          difficulty: new Difficulty(DIFFICULTY_KEY.EASY),
-          createdAt: new Date(),
-          materials: [
-            {
-              name: new Material('卵'),
-              count: 1
-            }
+          images: [
+            new PostImage('data:image/png;base64,iVBOR'),
+            new PostImage('data:image/png;base64,iVBOR'),
+            new PostImage('data:image/png;base64,iVBOR'),
+            new PostImage('data:image/png;base64,iVBOR'),
+            new PostImage('data:image/png;base64,iVBOR'),
+            new PostImage('data:image/png;base64,iVBOR')
           ],
-          processes: [],
-          userId: userId,
-          userStatus: new UserStatus(USER_STATUS_KEY.ACTIVE),
-          likes: null
-        })
-      ).toThrow('プロセスが設定されていません postId: ' + postId.get())
-    })
-    it('If the materials does not exist', () => {
-      expect(() =>
-        Post.create({
-          postId: postId,
-          title: new PostTitle('卵焼き'),
-          detail: new PostDetail('ちょー簡単な料理です。'),
           calories: new Calories(120),
           difficulty: new Difficulty(DIFFICULTY_KEY.EASY),
           createdAt: new Date(),
           materials: [],
-          processes: [
-            {
-              detail: new ProcessDetail('フライパンを熱して卵を入れます。'),
-              image: 'material.png'
-            }
-          ],
           userId: userId,
-          userStatus: new UserStatus(USER_STATUS_KEY.ACTIVE),
           likes: null
         })
       ).toThrow('材料が指定されていません postId: ' + postId.get())
@@ -146,6 +103,7 @@ describe('Post', () => {
         Post.create({
           postId: postId,
           title: new PostTitle('卵焼き'),
+          images: null,
           detail: new PostDetail('ちょー簡単な料理です。'),
           calories: new Calories(120),
           difficulty: new Difficulty(DIFFICULTY_KEY.EASY),
@@ -156,14 +114,7 @@ describe('Post', () => {
               count: 1
             }
           ],
-          processes: [
-            {
-              detail: new ProcessDetail('フライパンを熱して卵を入れます。'),
-              image: 'material.png'
-            }
-          ],
           userId: userId,
-          userStatus: new UserStatus(USER_STATUS_KEY.ACTIVE),
           likes: [userId]
         })
       ).toThrow('自分の投稿にいいねは出来ません userId: ' + userId.get())
@@ -176,6 +127,7 @@ describe('Post', () => {
           postId: postId,
           title: new PostTitle('卵焼き'),
           detail: new PostDetail('ちょー簡単な料理です。'),
+          images: null,
           calories: new Calories(120),
           difficulty: new Difficulty(DIFFICULTY_KEY.EASY),
           createdAt: new Date(),
@@ -185,44 +137,10 @@ describe('Post', () => {
               count: 1
             }
           ],
-          processes: [
-            {
-              detail: new ProcessDetail('フライパンを熱して卵を入れます。'),
-              image: 'material.png'
-            }
-          ],
           userId: userId,
-          userStatus: new UserStatus(USER_STATUS_KEY.ACTIVE),
           likes: [sameId, sameId, new UserId(ulid()), new UserId(ulid()), new UserId(ulid())]
         })
       ).toThrow('同一ユーザーに２回以上いいねされています userId: ' + userId.get())
-    })
-    it('If the user who submitted the post is not authenticated', () => {
-      expect(() =>
-        Post.create({
-          postId: postId,
-          title: new PostTitle('卵焼き'),
-          detail: new PostDetail('ちょー簡単な料理です。'),
-          calories: new Calories(120),
-          difficulty: new Difficulty(DIFFICULTY_KEY.EASY),
-          createdAt: new Date(),
-          materials: [
-            {
-              name: new Material('卵'),
-              count: 1
-            }
-          ],
-          processes: [
-            {
-              detail: new ProcessDetail('フライパンを熱して卵を入れます。'),
-              image: 'material.png'
-            }
-          ],
-          userId: userId,
-          userStatus: new UserStatus(USER_STATUS_KEY.PENDING),
-          likes: null
-        })
-      ).toThrow('登録済みのユーザー以外は投稿はできません')
     })
   })
 })
