+++ apps/domain/test/entity/User.test.ts
@@ -1,10 +1,8 @@
-import { USER_STATUS_KEY } from '@recipeaceful/library/dist/const'
 import { User } from '../../src/entity/User'
 import { ulid } from 'ulid'
 import { MailAddress } from '../../src/valueObject/MailAddress'
 import { PostId, UserId } from '../../src/valueObject/Ulid'
 import { UserName } from '../../src/valueObject/UserName'
-import { UserStatus } from '../../src/valueObject/UserStatus'
 
 describe('User', () => {
   const userId = new UserId(ulid())
@@ -17,7 +15,6 @@ describe('User', () => {
         email: mailAddress,
         password: '123456',
         icon: 'icon.png',
-        status: new UserStatus(USER_STATUS_KEY.ACTIVE),
         follows: [new UserId(ulid()), new UserId(ulid())],
         followers: [new UserId(ulid())],
         posts: [new PostId(ulid())]
@@ -28,7 +25,6 @@ describe('User', () => {
       expect(entity.email.get()).toBe('test@example.com')
       expect(entity.password).toBe('123456')
       expect(entity.icon).toBe('icon.png')
-      expect(entity.status.get()).toBe(USER_STATUS_KEY.ACTIVE)
       expect(entity.follows?.length).toBe(2)
       expect(entity.followers?.length).toBe(1)
       expect(entity.posts?.length).toBe(1)
@@ -40,7 +36,6 @@ describe('User', () => {
         email: mailAddress,
         password: '123456',
         icon: null,
-        status: new UserStatus(USER_STATUS_KEY.ACTIVE),
         follows: null,
         followers: null,
         posts: null
@@ -62,7 +57,6 @@ describe('User', () => {
           email: mailAddress,
           password: '123456',
           icon: 'icon.png',
-          status: new UserStatus(USER_STATUS_KEY.ACTIVE),
           follows: [new UserId(ulid()), new UserId(ulid())],
           followers: [userId],
           posts: [new PostId(ulid())]
@@ -77,7 +71,6 @@ describe('User', () => {
           email: mailAddress,
           password: '123456',
           icon: 'icon.png',
-          status: new UserStatus(USER_STATUS_KEY.ACTIVE),
           follows: [userId],
           followers: [new UserId(ulid())],
           posts: [new PostId(ulid())]
@@ -92,7 +85,6 @@ describe('User', () => {
           email: mailAddress,
           password: '123456',
           icon: 'icon.png',
-          status: new UserStatus(USER_STATUS_KEY.ACTIVE),
           follows: [new UserId(ulid())],
           followers: [sameId, sameId],
           posts: [new PostId(ulid())]
@@ -107,42 +99,11 @@ describe('User', () => {
           email: mailAddress,
           password: '123456',
           icon: 'icon.png',
-          status: new UserStatus(USER_STATUS_KEY.ACTIVE),
           follows: [sameId, sameId],
           followers: [new UserId(ulid())],
           posts: [new PostId(ulid())]
         })
       ).toThrow(`同一ユーザーを２回以上フォローしています userId: ${userId.get()}`)
-    })
-    it('if the user is not activated but the post does exist', () => {
-      expect(() =>
-        User.create({
-          userId: userId,
-          name: new UserName('テスト 太郎'),
-          email: mailAddress,
-          password: '123456',
-          icon: 'icon.png',
-          status: new UserStatus(USER_STATUS_KEY.PENDING),
-          follows: null,
-          followers: [new UserId(ulid())],
-          posts: [new PostId(ulid())]
-        })
-      ).toThrow(`仮登録中は投稿できません`)
-    })
-    it('if the user is not activated but the follows list does exist', () => {
-      expect(() =>
-        User.create({
-          userId: userId,
-          name: new UserName('テスト 太郎'),
-          email: mailAddress,
-          password: '123456',
-          icon: 'icon.png',
-          status: new UserStatus(USER_STATUS_KEY.PENDING),
-          follows: [new UserId(ulid())],
-          followers: [new UserId(ulid())],
-          posts: null
-        })
-      ).toThrow(`仮登録中はユーザーをフォローできません`)
     })
   })
 })
